@@ -2,12 +2,17 @@
     "name": "ride-saver",
     "version": "0.0.1-alpha",
     "devDependencies": {
-        "@babel/core": "^7.19.6",
+        "@babel/core": "^7.20.2",
         "@babel/preset-typescript": "^7.18.6",
-        "eslint": "^8.24.0",
+        "@typescript-eslint/eslint-plugin": "^5.43.0",
+        "@typescript-eslint/parser": "^5.43.0",
+        "eslint": "^8.27.0",
+        "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-prettier": "^4.2.1",
-        "gitlab-ci-local": "^4.33.1",
-        "jest": "^29.0.3",
+        "gitlab-ci-local": "^4.35.0",
+        "jest": "^29.3.1",
+        "jest-watch-select-projects": "^2.0.0",
+        "jest-watch-typeahead": "^2.2.0",
         "prettier": "^2.7.1",
         "prettier-eslint": "^15.0.1",
         "ts-jest": "^29.0.3",
@@ -15,7 +20,7 @@
     },
     "scripts": {
         "test:ci": "gitlab-ci-local --no-artifacts-to-source",
-        "test": "jest"
+        "test": "yarn node --experimental-vm-modules $(yarn bin jest)"
     },
     "workspaces": [
         "packages/*"
