@@ -10,30 +10,30 @@
   "dependencies": {
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.1",
-    "@react-navigation/stack": "^6.3.3",
+    "@react-navigation/stack": "^6.3.4",
     "@ridesaver/components": "0.0.1",
     "@ridesaver/internationalization": "0.0.1",
     "@ridesaver/store": "0.0.1",
-    "expo": "~46.0.9",
-    "expo-status-bar": "~1.4.0",
+    "expo": "~47.0.3",
+    "expo-status-bar": "~1.4.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-native": "0.70.4",
+    "react-native": "0.70.6",
     "react-native-gesture-handler": "~2.8.0",
     "react-native-paper": "^4.12.5",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.2",
     "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.10",
-    "react-redux": "^8.0.4"
+    "react-redux": "^8.0.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.6",
-    "@babel/preset-env": "^7.19.4",
-    "@expo/config": "^7.0.1",
-    "@types/react": "~18.0.14",
+    "@babel/core": "^7.20.2",
+    "@babel/preset-env": "^7.20.2",
+    "@expo/config": "^7.0.3",
+    "@types/react": "~18.0.25",
     "@types/react-native": "~0.70.6",
-    "typescript": "~4.8.4"
+    "typescript": "~4.9.3"
   },
   "private": true,
   "workspaces": {
