@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "@react-navigation/core": "^6.4.0",
-    "@react-navigation/stack": "^6.3.3",
+    "@react-navigation/stack": "^6.3.4",
     "@ridesaver/internationalization": "0.0.1",
     "@ridesaver/screens": "0.0.1",
     "@ridesaver/store": "0.0.1",
@@ -15,14 +15,14 @@
     "web-vitals": "^3.0.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
+    "@babel/core": "^7.20.2",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
     "@svgr/webpack": "^6.5.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
-    "babel-jest": "^29.2.2",
-    "babel-loader": "^9.0.1",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
     "babel-plugin-named-asset-import": "^0.3.8",
     "babel-preset-react-app": "^10.0.1",
     "bfj": "^7.0.2",
@@ -33,15 +33,16 @@
     "css-minimizer-webpack-plugin": "^4.2.2",
     "dotenv": "^16.0.3",
     "dotenv-expand": "^9.0.0",
-    "eslint": "^8.3.0",
+    "eslint": "^8.27.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-webpack-plugin": "^3.1.1",
     "file-loader": "^6.2.0",
     "fs-extra": "^10.0.0",
     "html-webpack-plugin": "^5.5.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.2.2",
-    "jest-resolve": "^29.2.2",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-resolve": "^29.3.1",
     "jest-watch-typeahead": "^2.2.0",
     "mini-css-extract-plugin": "^2.4.5",
     "postcss": "^8.4.4",
@@ -54,13 +55,13 @@
     "react-refresh": "^0.14.0",
     "resolve": "^1.20.0",
     "resolve-url-loader": "^5.0.0",
-    "sass-loader": "^13.1.0",
+    "sass-loader": "^13.2.0",
     "semver": "^7.3.5",
     "source-map-loader": "^4.0.1",
     "style-loader": "^3.3.1",
-    "tailwindcss": "^3.0.2",
+    "tailwindcss": "^3.2.3",
     "terser-webpack-plugin": "^5.2.5",
-    "webpack": "^5.64.4",
+    "webpack": "^5.75.0",
     "webpack-dev-server": "^4.6.0",
     "webpack-manifest-plugin": "^5.0.0",
     "workbox-webpack-plugin": "^6.4.1"
