@@ -1,8 +1,19 @@
-import { jest, describe, expect, test } from "@jest/globals";
-import { configureStore } from "@reduxjs/toolkit";
-import i18n, { locales, loadLocale as i18nLoadLocale, downloadMessages } from "@ridesaver/internationalization";
+/**
+ * Tests language loading handlers of redux state.
+ * @author Elias Schablowski
+ * @format
+ */
 
-const { slice, loadLocale } = require("..") as typeof import("..");
+import { jest, describe, expect, test } from "@jest/globals";
+import type { Mocked } from "jest-mock";
+import { configureStore } from "@reduxjs/toolkit";
+import i18n, {
+    locales,
+    loadLocale as i18nLoadLocale,
+    downloadMessages,
+} from "@ridesaver/internationalization";
+
+const { slice, loadLocale } = jest.requireActual("..") as typeof import("..");
 
 function setupStore() {
     return configureStore({
@@ -17,12 +28,16 @@ describe("loads language", () => {
         // Some dummy messages that should NEVER be in the real application
         const msgs = {
             cagfd: "ajfklds",
-            cagfds: "ajfklfdsads"
+            cagfds: "ajfklfdsads",
         };
-        (downloadMessages as any).mockImplementationOnce((locale)=>locale === "en_PS" ? msgs : {});
+        (
+            downloadMessages as Mocked<typeof downloadMessages>
+        ).mockImplementationOnce((locale) =>
+            Promise.resolve(locale === "en-PS" ? msgs : {})
+        );
         expect.assertions(1);
         const store = setupStore();
-        const messages = await store.dispatch(loadLocale("en_PS")).unwrap();
+        const messages = await store.dispatch(loadLocale("en-PS")).unwrap();
         expect(messages).toBe(msgs);
     });
     test("loads all possible locales correctly", async () => {
@@ -30,7 +45,7 @@ describe("loads language", () => {
 
         expect.assertions(3 * locales.length);
         for (const locale of locales) {
-            await store.dispatch(loadLocale(locale as any)).unwrap();
+            await store.dispatch(loadLocale(locale)).unwrap();
             const state = store.getState().language;
 
             expect(state.messages[locale]).toBeDefined();
