@@ -1,5 +1,8 @@
+/* eslint-env node */
+
 module.exports = {
-    "presets": ["@babel/preset-typescript", "@babel/preset-react"]
+    presets: ["@babel/preset-typescript", "@babel/preset-react"],
+    plugins: ["macros"],
     // root: __dirname,
     // rootMode: "root"
-}
+};
