@@ -1,19 +1,33 @@
+/**
+ * TextInput with specilizations for password input.
+ * @author Elias Schablowski
+ * @format
+ */
+
 import React, { useState } from "react";
 import { View } from "react-native";
-import { TextInput } from "react-native-paper";
+import { TextInput, TextInputProps } from "react-native-paper";
 import PasswordStrengthIndicator from "./PasswordStrengthIndicator";
 import { t } from "@lingui/macro";
 import i18n from "@ridesaver/internationalization";
 
 export default ({
-    showStrength,
+    password = "",
     onPasswordChange,
+    showStrength = false,
+    visible = false,
+    // eslint-disable-next-line @typescript-eslint/no-empty-function
+    onChangeVisibility = () => {},
+    ...options
 }: {
-    showStrength: boolean;
+    showStrength?: boolean;
+    password?: string;
     onPasswordChange: (password: string) => void;
-}) => {
-    const [password, setPassword] = useState("");
-    const [secure, setSecure] = React.useState(true);
+    visible?: boolean;
+    onChangeVisibility?: (visibility: boolean) => void;
+} & Partial<TextInputProps>) => {
+    const [_password, setPassword] = useState(password);
+    const [secure, setSecure] = React.useState(!visible);
     return (
         <View>
             <TextInput
@@ -23,18 +37,26 @@ export default ({
                 right={
                     <TextInput.Icon
                         icon="eye"
-                        onPress={() => setSecure(!secure)}
+                        onPress={() => {
+                            setSecure(!secure);
+                            onChangeVisibility(secure);
+                        }}
+                        testID="password-visibility-toggle"
                     />
                 }
-                value={password}
+                value={_password}
                 onChangeText={(text) => {
                     setPassword(text);
                     onPasswordChange(text);
                 }}
                 accessibilityLabel="Password"
+                {...options}
             />
             {showStrength ? (
-                <PasswordStrengthIndicator password={password} />
+                <PasswordStrengthIndicator
+                    password={_password}
+                    testID="password-strength"
+                />
             ) : undefined}
         </View>
     );
