@@ -1,4 +1,12 @@
+/**
+ * Reexports all components for easier access through module name.
+ * @author Elias Schablowski
+ * @format
+ */
+
 export { default as LocationSelector } from "./LocationSelector";
-export { default as RideList } from "./RideList";
-export { default as RideSettings } from "./RideSettings";
+export { default as RideEstimate } from "./RideEstimate";
+export { default as NumberInput } from "./NumberInput";
 export { default as PasswordInput } from "./PasswordInput";
+
+export { default as useDebounce } from "./utils/useDebounce";
