@@ -1,34 +1,44 @@
+/**
+ * The Home Screen for RideSaver.
+ * @author Elias Schablowski
+ * @format
+ */
+
 import React, { useCallback, useState } from "react";
 import { View } from "react-native";
-import { Button, HelperText, TextInput } from "react-native-paper";
-import { t, Trans } from "@lingui/macro";
-import { useLinkTo } from "@react-navigation/native";
+import { t } from "@lingui/macro";
 import { user, useDispatch } from "@ridesaver/store";
-import i18n from "@ridesaver/internationalization";
 import {
-  LocationSelector,
-  RideList,
-  RideSettings,
+    LocationSelector,
+    NumberInput,
+    RideEstimate,
 } from "@ridesaver/components";
-
+import type { Estimate } from "@RideSaver/api";
 
 export default () => {
     const dispatch = useDispatch();
-    const linkTo = useLinkTo();
 
     const [startPoint, setStartPoint] = useState(undefined);
     const [endPoint, setEndPoint] = useState(undefined);
+    const [riders, setRiders] = useState(1);
+    const estimates: Estimate[] = [];
 
     useCallback(() => {
         dispatch(user.load());
     }, []);
 
     return (
-      <View>
-        <LocationSelector onUpdateLocation={location=>setStartPoint} />
-        <LocationSelector onUpdateLocation={location=>setEndPoint} />
-        <RideSettings />
-        <RideList estimates={[]}/>
-      </View>
+        <View>
+            <LocationSelector onUpdateLocation={setStartPoint} />
+            <LocationSelector onUpdateLocation={setEndPoint} />
+            <NumberInput
+                value={riders}
+                onChangeValue={setRiders}
+                label={t`Riders`}
+            />
+            {estimates.map((estimate) => (
+                <RideEstimate estimate={estimate} key={estimate.id} />
+            ))}
+        </View>
     );
 };
