@@ -1,3 +1,9 @@
+/**
+ * Reexports store, selectors and actions for easy access, also correctly types useDispatch and useSelector for tighter typescript integration.
+ * @author Elias Schablowski
+ * @format
+ */
+
 import store, { RootState, AppDispatch } from "./store";
 import {
     useDispatch as reduxUseDispatch,
@@ -6,12 +12,11 @@ import {
 
 export { default as store } from "./store";
 export type { RootState, AppDispatch } from "./store";
-export * as rideSettings from "./slices/rideSettingsSelector";
 export * as language from "./slices/language";
 export * as user from "./slices/user";
 
 export const dispatch = store.dispatch;
-export const select = store.dispatch;
+export const getState = store.getState;
 
 export const useDispatch = reduxUseDispatch<AppDispatch>;
-export const useSelector = reduxUseSelector<AppDispatch>;
+export const useSelector = reduxUseSelector<RootState>;
