@@ -1,4 +1,10 @@
-import { jest, describe, expect, test, beforeEach } from "@jest/globals";
+/**
+ * Tests language switching of redux state.
+ * @author Elias Schablowski
+ * @format
+ */
+
+import { jest, describe, expect, test } from "@jest/globals";
 import type { Mocked } from "jest-mock";
 import { configureStore } from "@reduxjs/toolkit";
 
@@ -7,10 +13,11 @@ import i18n, {
     loadLocale as i18nLoadLocale,
 } from "@ridesaver/internationalization";
 
-const { slice, loadLocale, switchLocale } =
-    require("..") as typeof import("..");
+const { slice, loadLocale, switchLocale } = jest.requireActual(
+    ".."
+) as typeof import("..");
 
-jest.useFakeTimers();
+// jest.useFakeTimers();
 
 function setupStore() {
     return configureStore({
@@ -23,39 +30,39 @@ function setupStore() {
 describe("switch language", () => {
     test("languages have different same test message", async () => {
         const store = setupStore();
-        await store.dispatch((switchLocale as any)("en_US"));
-        expect(i18n.activate).toHaveBeenLastCalledWith("en_US");
+        await store.dispatch(switchLocale("en-US")).unwrap();
+        expect(i18n.activate).toHaveBeenLastCalledWith("en-US");
     });
 
     test("languages that were loaded should not be loaded again.", async () => {
         const store = setupStore();
-        (i18nLoadLocale as any).mockClear(); // Verify that mock is cleared
+        (i18nLoadLocale as Mocked<typeof i18nLoadLocale>).mockClear(); // Verify that mock is cleared
 
-        await store.dispatch(loadLocale("en_US"));
-        await store.dispatch((switchLocale as any)("en_US"));
+        await store.dispatch(loadLocale("en-US"));
+        await store.dispatch(switchLocale("en-US"));
 
         /**
          * Expect that the locale was loaded
          */
         expect(i18nLoadLocale).toHaveBeenCalledTimes(1); // should only be called once for a preloaded locale
 
-        await store.dispatch(loadLocale("en_UK"));
-        await store.dispatch((switchLocale as any)("en_UK"));
+        await store.dispatch(loadLocale("en-UK"));
+        await store.dispatch(switchLocale("en-UK"));
 
         expect(i18nLoadLocale).toHaveBeenCalledTimes(2); // Should be called a second time for a different locale
     });
 
     test("languages that were not loaded should be loaded.", async () => {
         const store = setupStore();
-        (i18nLoadLocale as any).mockClear(); // Verify that mock is cleared
+        (i18nLoadLocale as Mocked<typeof i18nLoadLocale>).mockClear(); // Verify that mock is cleared
 
-        await store.dispatch((switchLocale as any)("en_US"));
+        await store.dispatch(switchLocale("en-US"));
 
         /**
          * Expect that the locale was loaded with the correct locale
          */
         expect(i18nLoadLocale).toHaveBeenCalledWith(
-            "en_US",
+            "en-US",
             expect.anything(),
             expect.anything()
         );
@@ -63,30 +70,30 @@ describe("switch language", () => {
 
     test("last selected language should be activated.", async () => {
         const store = setupStore();
-        (i18nLoadLocale as any).mockClear(); // Verify that mock is cleared
+        (i18nLoadLocale as Mocked<typeof i18nLoadLocale>).mockClear(); // Verify that mock is cleared
 
         (
             downloadMessages as Mocked<typeof downloadMessages>
         ).mockImplementationOnce(
             (locale) =>
-                new Promise((resolve, reject) => {
+                new Promise((resolve) => {
                     setTimeout(
                         () =>
                             resolve({
                                 locale,
                             }),
-                        10000
+                        10
                     );
                 })
         );
-        store.dispatch((switchLocale as any)("en_US"));
-        await store.dispatch((switchLocale as any)("en_UK"));
+        const switch1 = store.dispatch(switchLocale("en-US")).unwrap();
+        await store.dispatch(switchLocale("en-UK"));
 
-        jest.runAllTimers();
-        
+        await switch1;
+
         /**
          * Expect that the locale was loaded with the correct locale
          */
-        expect(i18n.activate).toHaveBeenLastCalledWith("en_UK");
+        expect(i18n.activate).toHaveBeenLastCalledWith("en-UK");
     });
 });
