@@ -1,12 +1,16 @@
+/**
+ * Create the Redux store, as well as extract typings from the store.
+ * @author Elias Schablowski
+ * @format
+ */
+
 import { configureStore } from "@reduxjs/toolkit";
 
-import { slice as rideSettingsSlice } from "./slices/rideSettingsSelector";
 import { slice as languageSlice } from "./slices/language";
 import { slice as userSlice } from "./slices/user";
 
 const store = configureStore({
     reducer: {
-        rideSettings: rideSettingsSlice.reducer,
         language: languageSlice.reducer,
         user: userSlice.reducer,
     },
