@@ -1,4 +1,4 @@
-'use strict';
+
 
 const babelJest = require('babel-jest').default;
 
