@@ -1,4 +1,4 @@
-'use strict';
+
 
 // This is a custom Jest transformer turning style imports into empty objects.
 // http://facebook.github.io/jest/docs/en/webpack.html
