@@ -1,4 +1,10 @@
-import React from "react";
+/**
+ * Provider and Screen configuration for the web based application
+ * @author Elias Schablowski
+ * @format
+ */
+
+import React, { useCallback } from "react";
 import { useColorScheme } from "react-native";
 import { Provider as ReduxProvider } from "react-redux";
 import * as Screens from "@ridesaver/screens";
@@ -14,12 +20,15 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { NavigationContainer } from "@react-navigation/native";
 
 import * as Themes from "./themes";
+import { language } from "@ridesaver/store";
 
 export default function App() {
     const state = store.getState();
     const scheme = useColorScheme();
     const theme = scheme === "dark" ? Themes.dark : Themes.light;
-    console.log(theme.colors)
+    useCallback(() => {
+        store.dispatch(language.switchLocale("en-US"));
+    }, []);
 
     if (state.user.isLoading) {
         // We haven't finished checking for the token yet
@@ -35,64 +44,64 @@ export default function App() {
             }}
             theme={theme}
         >
-                <I18nProvider i18n={i18n}>
-                    <ReduxProvider store={store}>
-                        <NavigationContainer
-                            linking={{
-                                prefixes: [
-                                    "ridesaver://",
-                                    ...["https://", "http://", "//", ""].map(
-                                        (protocol) =>
-                                            protocol + window.location.host
-                                    ),
-                                ],
-                                config: {
-                                    screens: Screens.paths,
-                                },
-                            }}
-                            documentTitle={{
-                                formatter: (options, route) =>
-                                    t(i18n)`${
-                                        options?.title ?? route?.name
-                                    } - RideSaver`,
-                            }}
-                            theme={theme}
-                        >
-                            <Stack.Navigator>
-                                {state.user.token === undefined ? (
-                                    // No token found, user isn't signed in
-                                    <>
-                                        <Stack.Screen
-                                            name="Login"
-                                            component={Screens.Login}
-                                            options={{
-                                                title: t(i18n)`Login`,
-                                            }}
-                                        />
-                                        <Stack.Screen
-                                            name="SignUp"
-                                            component={Screens.SignUp}
-                                            options={{
-                                                title: t(i18n)`Sign Up`,
-                                            }}
-                                        />
-                                    </>
-                                ) : (
-                                    // User is signed in
-                                    <>
-                                        <Stack.Screen
-                                            name="Home"
-                                            component={Screens.Home}
-                                            options={{
-                                                title: t(i18n)`Home`,
-                                            }}
-                                        />
-                                    </>
-                                )}
-                            </Stack.Navigator>
-                        </NavigationContainer>
-                    </ReduxProvider>
-                </I18nProvider>
+            <I18nProvider i18n={i18n}>
+                <ReduxProvider store={store}>
+                    <NavigationContainer
+                        linking={{
+                            prefixes: [
+                                "ridesaver://",
+                                ...["https://", "http://", "//", ""].map(
+                                    (protocol) =>
+                                        protocol + window.location.host
+                                ),
+                            ],
+                            config: {
+                                screens: Screens.paths,
+                            },
+                        }}
+                        documentTitle={{
+                            formatter: (options, route) =>
+                                t(i18n)`${
+                                    options?.title ?? route?.name
+                                } - RideSaver`,
+                        }}
+                        theme={theme}
+                    >
+                        <Stack.Navigator>
+                            {state.user.token === undefined ? (
+                                // No token found, user isn't signed in
+                                <>
+                                    <Stack.Screen
+                                        name="Login"
+                                        component={Screens.Login}
+                                        options={{
+                                            title: t(i18n)`Login`,
+                                        }}
+                                    />
+                                    <Stack.Screen
+                                        name="SignUp"
+                                        component={Screens.SignUp}
+                                        options={{
+                                            title: t(i18n)`Sign Up`,
+                                        }}
+                                    />
+                                </>
+                            ) : (
+                                // User is signed in
+                                <>
+                                    <Stack.Screen
+                                        name="Home"
+                                        component={Screens.Home}
+                                        options={{
+                                            title: t(i18n)`Home`,
+                                        }}
+                                    />
+                                </>
+                            )}
+                        </Stack.Navigator>
+                    </NavigationContainer>
+                </ReduxProvider>
+            </I18nProvider>
             <style type="text/css">
                 {`@font-face {
                     font-family: 'MaterialCommunityIcons';
