@@ -1,8 +1,17 @@
+/**
+ * Render the web based app on the browser. Also supports hydration.
+ * @author Elias Schablowski
+ * @format
+ */
+
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
+/**
+ * @brief
+ */
 const rootNode = document.getElementById("root");
 let root: ReactDOM.Root;
 const elem = (
@@ -15,7 +24,7 @@ if (rootNode.childElementCount > 0) {
 } else {
     root = ReactDOM.createRoot(document.getElementById("root"));
 }
-root.render(elem);
+root && root.render(elem);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
